rygg/integration_tests/requirements.txt
pytest>=6.2.4
python-dateutil>=2.8.2
requests>=2.26
retrying>=1.3.3
